Change dependency from docker-py to docker (name of the python API for docker)
docker seems to be the new python module name for the python docker API,
replacing docker-py.

The issue fixed in the previous commit is that the "docker" module (v
2.5.1) was installed when I installed docker-squash (as a dep), beside
docker-py (v 1.10.6).

Afterward, oardocker then uses the docker module instead of docker-py,
which cause the errors.

It seems that oardocker should then depends on the docker module, not
docker-py, if we want to avoid the comflict.

This is what this commit fixes. Hope this does not break the rest of the
code...